Run bitbake with -k to continue as far as possible
If an error occurs it's best to continue building as far as possible so
that other errors can also be exposed. If the build stops on the first
error you won't find out there are more until you fix the first.
This will also leave the sstate in a more usable state when an error is
hit.

Signed-off-by: George McCollister <george.mccollister@gmail.com>